make mpsse_contect opac, add private header, remove any ref to ftid1 from public header
as mpsse.h is installed as a public header, it should not contain
conditional include of ftdi1.h as LIBFTDI1 will be undefined even when
the library was built with it defined.
for a library user, the content of the struct mpsse_context is
irrelevant, so make it a opac pointer.
adjust all internal includes to also include the private header